Conflicts:
	3rdParty/Src/mvvmlight_a9071d84ec65/GalaSoft.MvvmLight/GalaSoft.MvvmLight.Extras (NET4)/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
	3rdParty/Src/mvvmlight_a9071d84ec65/GalaSoft.MvvmLight/GalaSoft.MvvmLight.Test (NET4)/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache

+++ Users/Zack/CloudSDK_SmokeTest/Managers/FileManager.cs
@@ -1,8 +1,8 @@
-﻿using CloudApiPublic;
-using CloudApiPublic.Interfaces;
-using CloudApiPublic.JsonContracts;
-using CloudApiPublic.Model;
-using CloudApiPublic.Static;
+﻿using Cloud;
+using Cloud.Interfaces;
+using Cloud.JsonContracts;
+using Cloud.Model;
+using Cloud.Static;
 using CloudSDK_SmokeTest.Events.CLEventArgs;
 using CloudSDK_SmokeTest.Events.ManagerEventArgs;
 using CloudSDK_SmokeTest.Helpers;
@@ -133,7 +133,7 @@ namespace CloudSDK_SmokeTest.Managers
             int createReturnCode = 0;
             List<FileChange> folderChanges = new List<FileChange>();
             fileChange = GetFileChange(isFile, eventArgs, ref folderChanges);
-            CloudApiPublic.JsonContracts.Event returnEvent;
+            Cloud.JsonContracts.Event returnEvent;
             CLHttpRestStatus restStatus = eventArgs.RestStatus;
 
             CLError postFolderError;
@@ -312,7 +312,7 @@ namespace CloudSDK_SmokeTest.Managers
         private int BeginManualRename(SmokeTestManagerEventArgs e, Rename renameTask)
         {
             FileChange fileChange = null;
-            CloudApiPublic.JsonContracts.Metadata metaData = new CloudApiPublic.JsonContracts.Metadata();
+            Cloud.JsonContracts.Metadata metaData = new Cloud.JsonContracts.Metadata();
             bool isFile = renameTask.ObjectType.type == ModificationObjectType.File;
             if ((isFile && e.FileInfo == null) || (!isFile && e.DirectoryInfo == null))
             {
@@ -352,7 +352,7 @@ namespace CloudSDK_SmokeTest.Managers
             StringBuilder newBuilder = new StringBuilder("Begining File Rename...");
             int responseCode = (int)FileManagerResponseCodes.Success;
             CLHttpRestStatus restStatus = e.RestStatus;
-            CloudApiPublic.JsonContracts.Event returnEvent = e.ReturnEvent;
+            Cloud.JsonContracts.Event returnEvent = e.ReturnEvent;
             CLError postFileError = e.SyncBox.HttpRestClient.PostFileChange(e.FileChange, ManagerConstants.TimeOutMilliseconds, out restStatus, out returnEvent);
             e.ReturnEvent = returnEvent;
             e.RestStatus = restStatus;
@@ -373,11 +373,47 @@ namespace CloudSDK_SmokeTest.Managers
                 try
                 {
                     if (renameTask.ObjectType.type == ModificationObjectType.Folder)
-                        Directory.Move(e.FileChange.OldPath.ToString(), e.FileChange.NewPath.ToString());
+                    {
+                        if (!Directory.Exists(e.FileChange.NewPath.ToString()))
+                        {
+                            Directory.Move(e.FileChange.OldPath.ToString(), e.FileChange.NewPath.ToString());
+                            if(Directory.Exists(e.FileChange.NewPath.ToString()))
+                            {
+                                newBuilder.AppendLine("Successfully Renamed Folder:");
+                                newBuilder.AppendLine(string.Format("  From: {0}", e.FileChange.OldPath.ToString()));
+                                newBuilder.AppendLine(string.Format("    To: {0}", e.FileChange.NewPath.ToString()));
+                            }
+                        }
+                        else
+                        {
+                            newBuilder.AppendLine("Failed to Rename Folder Because Source Already Exists: ");
+                            newBuilder.AppendLine(string.Format("     Source Folder      : {0}", e.FileChange.OldPath.ToString()));
+                            newBuilder.AppendLine(string.Format("     Destination Folder : {0}", e.FileChange.NewPath.ToString()));
+                            newBuilder.AppendLine();
+                        }
+                    }
                     else if (renameTask.ObjectType.type == ModificationObjectType.File)
-                        File.Move(e.FileChange.OldPath.ToString(), e.FileChange.NewPath.ToString());
+                    {
+                        if (!File.Exists(e.FileChange.NewPath.ToString()))
+                        {
+                            File.Move(e.FileChange.OldPath.ToString(), e.FileChange.NewPath.ToString());
+                            if(File.Exists(e.FileChange.NewPath.ToString()))
+                            {
+                                newBuilder.AppendLine("Successfully Renamed File:");
+                                newBuilder.AppendLine(string.Format("   From: {0}", e.FileChange.OldPath.ToString()));
+                                newBuilder.AppendLine(string.Format("     To: {0}", e.FileChange.NewPath.ToString()));
+                            }
+                        }
+                        else
+                        {
+                            newBuilder.AppendLine("Failed to Rename File Because Source Already Exists: ");
+                            newBuilder.AppendLine(string.Format("     Source File      : {0}", e.FileChange.OldPath.ToString()));
+                            newBuilder.AppendLine(string.Format("     Destination File : {0}", e.FileChange.NewPath.ToString()));
+                            newBuilder.AppendLine();
+                        }
+                    }
 
-                    newBuilder.AppendLine(string.Format("Successfully Renamed File {0} to {1}", e.FileChange.OldPath.ToString(), e.FileChange.NewPath.ToString()));
+                   
                 }
                 catch (Exception excetpion)
                 {
@@ -458,6 +494,9 @@ namespace CloudSDK_SmokeTest.Managers
             try
             {
                 File.Move(oldPath, newPath);
+                FileInfo newInfo = new FileInfo(newPath);
+                if(!newInfo.Exists)
+                    throw new Exception(string.Format("{0} Not Moved to {1}.", oldPath, newPath));
                 reportBuilder.AppendLine();
                 reportBuilder.AppendLine("Successfully Renamed Item:");
                 reportBuilder.AppendLine(string.Format("    From: {0}", oldPath));
@@ -606,9 +645,12 @@ namespace CloudSDK_SmokeTest.Managers
                 rootFolder = TrimTrailingSlash(e.ParamSet.ManualSync_Folder.Replace("\"", ""));
             else
                 rootFolder = TrimTrailingSlash(e.ParamSet.ActiveSync_Folder.Replace("\"", ""));
-                
-                
-            relative = TrimTrailingSlash(renameTask.RelativeDirectoryPath);
+
+            if (FileHelper.PathEndsWithSlash(renameTask.RelativeDirectoryPath) && renameTask.RelativeDirectoryPath.Count() > 1)
+                relative = TrimTrailingSlash(renameTask.RelativeDirectoryPath);
+            else
+                relative = renameTask.RelativeDirectoryPath;
+
             fullPath = string.Concat(rootFolder, relative, renameTask.OldName.Replace("\"", ""));
             if (Directory.Exists(fullPath))
                 returnInfo = new DirectoryInfo(fullPath);
@@ -739,7 +781,7 @@ namespace CloudSDK_SmokeTest.Managers
             else
                 metadataList = GetFoldersForDelete(e, e.CurrentTask as Deletion).ToList();
             List<FileChange> changes = GetFileChangesFromMetadata(e.SyncBox, metadataList).ToList();
-            CloudApiPublic.JsonContracts.Event returnEvent;
+            Cloud.JsonContracts.Event returnEvent;
 
             foreach (FileChange fc in changes)
             {
@@ -818,7 +860,7 @@ namespace CloudSDK_SmokeTest.Managers
             List<Metadata> returnValues = new List<Metadata>();
             GenericHolder<CLError> refHolder = folderDeleteArgs.ProcessingErrorHolder;
             CLHttpRestStatus restStatus;
-            CloudApiPublic.JsonContracts.Folders folders;
+            Cloud.JsonContracts.Folders folders;
             CLError getFilesError = folderDeleteArgs.SyncBox.GetFolderHierarchy(ManagerConstants.TimeOutMilliseconds, out restStatus, out folders);
             if (getFilesError != null || restStatus != CLHttpRestStatus.Success)
             {
@@ -912,7 +954,7 @@ namespace CloudSDK_SmokeTest.Managers
             {
                 if (!Directory.Exists(eventArgs.DirectoryInfo.FullName))
                     Directory.CreateDirectory(eventArgs.DirectoryInfo.FullName);
-                CloudApiPublic.JsonContracts.Metadata metadata = new CloudApiPublic.JsonContracts.Metadata()
+                Cloud.JsonContracts.Metadata metadata = new Cloud.JsonContracts.Metadata()
                 {
                     CreatedDate = DateTime.UtcNow,
                     IsFolder = true,
@@ -922,7 +964,7 @@ namespace CloudSDK_SmokeTest.Managers
             }
         }
 
-        private FileChange PrepareFileChangeForModification(SmokeTestManagerEventArgs e, FileChangeType changeType, CloudApiPublic.JsonContracts.Metadata metaDataResponse)
+        private FileChange PrepareFileChangeForModification(SmokeTestManagerEventArgs e, FileChangeType changeType, Cloud.JsonContracts.Metadata metaDataResponse)
         {
 
             FileChange fileChange = new FileChange();
@@ -939,7 +981,7 @@ namespace CloudSDK_SmokeTest.Managers
 
             try
             {
-                CLError badPathError = CloudApiPublic.Static.Helpers.CheckForBadPath(e.FileInfo.FullName);
+                CLError badPathError = Cloud.Static.Helpers.CheckForBadPath(e.FileInfo.FullName);
                 //TODO: Add call to check files exists   /// try syncbox.GetVersions 
                 if (metaDataResponse.Deleted == null || metaDataResponse.Deleted == false)
                     md5Bytes = FileHelper.CreateFileChangeObject(e.FileInfo.FullName, changeType, true, metaDataResponse.Size, metaDataResponse.StorageKey, metaDataResponse.ServerId, out fileChange);
@@ -955,7 +997,7 @@ namespace CloudSDK_SmokeTest.Managers
             return fileChange;
         }
 
-        private bool GetMetadata(SmokeTestManagerEventArgs e, bool isFile, out CloudApiPublic.JsonContracts.Metadata metaData)
+        private bool GetMetadata(SmokeTestManagerEventArgs e, bool isFile, out Cloud.JsonContracts.Metadata metaData)
         {
             bool success = true;
             CLHttpRestStatus restStatus = CLHttpRestStatus.BadRequest;
@@ -1052,7 +1094,7 @@ namespace CloudSDK_SmokeTest.Managers
             return fileChanges;
         }
 
-        private FileMetadata GetFileMetadataFromMetadata(CloudApiPublic.JsonContracts.Metadata item, bool isFolder)
+        private FileMetadata GetFileMetadataFromMetadata(Cloud.JsonContracts.Metadata item, bool isFolder)
         {
             FileMetadata metadata = new FileMetadata();
             metadata.HashableProperties = new FileMetadataHashableProperties(isFolder, item.ModifiedDate.Value, item.CreatedDate.Value, item.Size);
